Bar Graph for Prior and Posterior Probability
Not sure why C and B labels are switched. Do python for-loops not
iterate through dictionaries in order of their keys? That needs to be
fixed somehow, and if I don't find out I'll switch their placements in
the original dictionary. Also not sure what I should do to make this bar
graph look nicer or more appropriate for a poster.